Default spi layer implementation

+++ gcloud-java-resourcemanager/src/main/java/com/google/gcloud/spi/ResourceManagerRpc.java
@@ -48,7 +48,7 @@ public interface ResourceManagerRpc {
       return get(options);
     }
 
-    Long getInt(Map<Option, ?> options) {
+    Integer getInt(Map<Option, ?> options) {
       return get(options);
     }
   }
